+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/binder/infopanel/InfoPanelTripProgressBinder.kt
@@ -5,13 +5,17 @@ import android.transition.TransitionManager
 import android.view.ViewGroup
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
+import com.mapbox.navigation.dropin.NavigationViewContext
 import com.mapbox.navigation.dropin.R
 import com.mapbox.navigation.dropin.binder.UIBinder
 import com.mapbox.navigation.dropin.component.tripprogress.TripProgressComponent
 import com.mapbox.navigation.dropin.databinding.MapboxInfoPanelTripProgressLayoutBinding
+import com.mapbox.navigation.dropin.internal.extensions.reloadOnChange
 
 @ExperimentalPreviewMapboxNavigationAPI
-internal class InfoPanelTripProgressBinder : UIBinder {
+internal class InfoPanelTripProgressBinder(
+    private val navigationViewContext: NavigationViewContext
+) : UIBinder {
 
     override fun bind(viewGroup: ViewGroup): MapboxNavigationObserver {
         val scene = Scene.getSceneForLayout(
@@ -22,6 +26,13 @@ internal class InfoPanelTripProgressBinder : UIBinder {
         TransitionManager.go(scene)
 
         val binding = MapboxInfoPanelTripProgressLayoutBinding.bind(viewGroup)
-        return TripProgressComponent(binding.tripProgressView)
+        return reloadOnChange(
+            navigationViewContext.styles.tripProgressStyle
+        ) { styles ->
+            TripProgressComponent(
+                styles = styles,
+                tripProgressView = binding.tripProgressView
+            )
+        }
     }
 }