+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/component/speedlimit/SpeedLimitComponent.kt
@@ -1,5 +1,6 @@
 package com.mapbox.navigation.dropin.component.speedlimit
 
+import androidx.annotation.StyleRes
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.dropin.internal.extensions.flowLocationMatcherResult
@@ -11,9 +12,16 @@ import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 @ExperimentalPreviewMapboxNavigationAPI
-internal class SpeedLimitComponent(val speedLimitView: MapboxSpeedLimitView) : UIComponent() {
+internal class SpeedLimitComponent(
+    @StyleRes val style: Int,
+    @StyleRes val textAppearance: Int,
+    val speedLimitView: MapboxSpeedLimitView
+) : UIComponent() {
     override fun onAttached(mapboxNavigation: MapboxNavigation) {
         super.onAttached(mapboxNavigation)
+        speedLimitView.updateStyle(style)
+        // setTextAppearance is not deprecated in AppCompatTextView
+        speedLimitView.setTextAppearance(speedLimitView.context, textAppearance)
         val speedLimitFormatter = SpeedLimitFormatter(speedLimitView.context)
         val speedLimitApi = MapboxSpeedLimitApi(speedLimitFormatter)